remove legacy header and footer
