--- assets/accordion-overrides.scss
@@ -1,7 +1,0 @@
-.main-content {
-  .el-collapse-item__header, .el-collapse-item__content {
-    font-size: 1rem;
-    padding-inline: 32px 16px;
-  }
-}
-
